+++ pages/index.js
@@ -1,16 +1,21 @@
 import React from "react";
 import Head from 'next/head'
 import Layout from "../components/Layout";
+import Banner from "../components/Banner";
 
-const Index = () => (
+const Index = () => {
+    const indexBanner  = "/static/banner.svg";
+
+    return (
     <Layout>
         <Head>
             <title>Home</title>
         </Head>
-        <div className={"bar-margin"}>
+        <Banner imageUrl={indexBanner}/>
+        <div className={"container"}>
             <h1>Hello Next.js</h1>
         </div>
-    </Layout>
-);
+    </Layout>);
+};
 
 export default Index;